backend/requirements.txt
Django>=3.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
redis>=5.0.1
elasticsearch>=8.11.0
tensorflow-recommenders>=0.7.3
transformers>=4.36.0
celery>=5.3.6
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
pika>=1.3.1
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
django-filter>=23.5
django-storages>=1.14.2
boto3>=1.34.0
Pillow>=10.1.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2 